musicplatform/pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = "django-training"
authors = ["Ahmed Mohamed Kamel <ahmedmohamedkamel06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-model-utils = "^4.2.0"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
python-dotenv = "^0.21.0"
django-imagekit = "^4.1.0"
Pillow = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
